+++ sickrage/core/databases/cache/__init__.py
@@ -25,6 +25,7 @@ from sickrage.core.common import Quality
 from sickrage.core.databases import srDatabase
 from sickrage.core.databases.cache.index import CacheLastUpdateIndex, CacheLastSearchIndex, CacheSceneExceptionsIndex, \
     CacheSceneNamesIndex, CacheNetworkTimezonesIndex, CacheSceneExceptionsRefreshIndex, CacheProvidersIndex
+from sickrage.core.helpers import validate_url
 
 
 class CacheDB(srDatabase):
@@ -57,4 +58,6 @@ class CacheDB(srDatabase):
     def cleanup_provider_cache(self):
         for item in self.all('providers'):
             if int(item["quality"]) == Quality.UNKNOWN:
+                self.delete(item)
+            elif not validate_url(item["url"]) and not item["url"].startswith("magnet"):
                 self.delete(item)